backend/pyproject.toml
[tool.poetry]
name = "NTAI API"
version = "0.1.0"
description = ""
authors = ["Ayush Likhar <ayush.likhar@sumasoft.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
sqlmodel = "^0.0.16"
pydantic = {extras = ["email"], version = "^2.7.0"}
slowapi = "^0.1.9"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
